+++ src/migrations/202510171723_tokens.ts
@@ -1,0 +1,13 @@
+import type { Knex } from 'knex';
+
+export const up = async (knex: Knex): Promise<void> => {
+  await knex.schema.alterTable('user', (table: Knex.TableBuilder) => {
+    table.specificType('tokens', 'character varying(255)[]');
+  });
+};
+
+export const down = async (knex: Knex): Promise<void> => {
+  await knex.schema.alterTable('user', (table: Knex.TableBuilder) => {
+    table.dropColumn('tokens');
+  });
+};