+++ src/types.ts
@@ -1,4 +1,5 @@
 import express from 'express';
+import { Knex } from 'knex';
 
 export type JsonPrimative = string | number | boolean | null;
 export type JsonArray = Json[];
@@ -9,6 +10,12 @@ export type Json = JsonPrimative | JsonComposite;
 export interface Model {
   toJSON: (req: Request) => any;
   getVocabulary: (req: Request) => any;
+}
+
+export interface User extends Model {
+  id: string;
+  tokens: string[];
+  update: (data: any, trx: Knex.Transaction) => Promise<void>;
 }
 
 export interface Request extends express.Request {
@@ -23,9 +30,7 @@ export interface Request extends express.Request {
     filter_content_types: boolean;
     allowed_content_types: string[];
   };
-  user: {
-    id: string;
-  };
+  user: User;
   token?: string;
   timestamp: string;
 }